// src/components/ui/practice/itr/itr-7/components/itr-7-schedule-80ie.types.ts
export interface ITR7Schedule80IEData {
    schedule80IE: {
        assam: { undertaking1: number; undertaking2: number };
        arunachal_pradesh: { undertaking1: number; undertaking2: number };
        manipur: { undertaking1: number; undertaking2: number };
        mizoram: { undertaking1: number; undertaking2: number };
        meghalaya: { undertaking1: number; undertaking2: number };
        nagaland: { undertaking1: number; undertaking2: number };
        tripura: { undertaking1: number; undertaking2: number };
        sikkim: { undertaking1: number; undertaking2: number };
        total_deduction_north_east: number; // ai
        total_deduction_80ie: number; // b
    };
}